refactor: move snippets to app dir
